+++ libEntendre/include/ConsecutiveNeuralNet.hh
@@ -10,14 +10,14 @@ class ConsecutiveNeuralNet : public NeuralNet<Node,Connection> {
 public:
   using NeuralNet::NeuralNet;
 
-  void load_input_vals(const std::vector<double>& inputs);
-  std::vector<double> read_output_vals();
-  std::vector<double> evaluate(std::vector<double> inputs) override;
+  void load_input_vals(const std::vector<float>& inputs);
+  std::vector<float> read_output_vals();
+  std::vector<float> evaluate(std::vector<float> inputs) override;
 
   std::vector<NodeType> node_types() const;
 protected:
-  double get_node_val(unsigned int i);
-  void add_to_val(unsigned int i, double val);
+  float get_node_val(unsigned int i);
+  void add_to_val(unsigned int i, float val);
 
 
 private:
